requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-decouple==3.8
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.5.0
